fix(rendering): add missing uint32_t cast

+++ engine/src/cubos/engine/renderer/plugin.cpp
@@ -65,7 +65,7 @@ static void frameGrids(const Assets& assets, Renderer& renderer, RendererFrame& 
             grid.handle = renderer->upload(gridRead.get());
         }
 
-        CUBOS_ASSERT(entity.index < (1 << 31), "Entity index must be lower than 2^31");
+        CUBOS_ASSERT(entity.index < static_cast<uint32_t>(1 << 31), "Entity index must be lower than 2^31");
         frame.draw(grid.handle, localToWorld.mat * glm::translate(glm::mat4(1.0F), grid.offset), entity.index);
     }
 }
